Assignment_3/code/requirements.txt
confluent-kafka
cassandra-driver
argparse
pandas
flask
pySpark